Moved names to plans and devices
