+++ nsls2tools/csx1/startup/endstation.py
@@ -2,10 +2,10 @@ from ophyd import (PVPositioner, EpicsMotor, EpicsSignal, EpicsSignalRO,
                    PVPositionerPC, Device)
 from ophyd import Component as Cpt
 
-from ..ophyd.optics import (SamplePosVirtualMotor, Cryoangle,
+from ..devices.optics import (SamplePosVirtualMotor, Cryoangle,
                              Nanopositioner)
 
-from ..ophyd.lakeshore import Lakeshore336
+from ..devices.lakeshore import Lakeshore336
 
 
 # Diffo angles
